Fixed delete func, added add button

+++ src/components/Carlist.jsx
@@ -1,11 +1,13 @@
-import {useState, useEffect, useRef} from 'react';
+import {useState, useEffect} from 'react';
 import {AgGridReact} from 'ag-grid-react';
 import 'ag-grid-community/styles/ag-grid.css';
 import 'ag-grid-community/styles/ag-theme-material.css';
+import Button from '@mui/material/Button';
+import ConfirmDelete from './ConfirmDelete';
+import AddCar from './AddCar';
 
 export default function Carlist() {
     const [cars, setCars] = useState([]);
-    const gridRef = useRef();
 
     useEffect(()=> fetchData(), []);
 
@@ -19,12 +21,26 @@ export default function Carlist() {
     const buttonRenderer = (row) => {
         var link = row.data._links.self.href
         return(
-            <button className='del-btn' onClick={() => handleDelete(link)}>Delete</button>
+            <Button variant="contained" size="small" color="error" onClick={() => handleDelete(link)}>Delete</Button>
         )
     };
 
     const handleDelete = (link) => {
-        fetch(link, {method: 'DELETE'})
+        if(window.confirm('Are you sure?')){
+            fetch(link, {method: 'DELETE'})
+            .then(response => fetchData())
+            .catch(err => console.error(err))
+        }
+    };
+
+    const saveCar = (car) => {
+        fetch('http://carrestapi.herokuapp.com/cars', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+            body: JSON.stringify(car)
+        })
         .then(response => fetchData())
         .catch(err => console.error(err))
     };
@@ -33,50 +49,66 @@ export default function Carlist() {
         {
             headerName: 'Brand',
             field: 'brand',
-            filter: 'true'
+            sortable: true,
+            filter: 'agTextColumnFilter',
+            suppressMenu: true,
+            floatingFilter: true
         },
         {
             headerName: 'Model',
             field: 'model',
-            filter: 'true'
+            sortable: true,
+            filter: 'agTextColumnFilter',
+            suppressMenu: true,
+            floatingFilter: true
         },
         {
             headerName: 'Color',
             field: 'color',
-            filter: 'true'
+            sortable: true,
+            filter: 'agTextColumnFilter',
+            suppressMenu: true,
+            floatingFilter: true
         },
         {
             headerName: 'Fuel',
             field: 'fuel',
-            filter: 'true'
+            sortable: true,
+            filter: 'agTextColumnFilter',
+            suppressMenu: true,
+            floatingFilter: true
         },
         {
             headerName: 'Year',
             field: 'year',
-            filter: 'true'
+            sortable: true,
+            filter: 'agNumberColumnFilter',
+            suppressMenu: true,
+            floatingFilter: true
         },
         {
             headerName: 'Price',
             field: 'price',
-            filter: 'true'
+            sortable: true,
+            filter: 'agNumberColumnFilter',
+            suppressMenu: true,
+            floatingFilter: true
         },
         {
-            width: '100px',
+            width: 100,
             cellRenderer: buttonRenderer
         }
     ]
 
     return(
         <div className='ag-theme-material' style={{width: '1350px', height: '600px', margin: 'auto', padding: '20px 0'}}>
+            <AddCar saveCar={saveCar}/>
             <AgGridReact 
                 columnDefs={columns}
                 rowData={cars}
-                ref={gridRef}
-                onGridReady={ params => gridRef.current = params.api }
-                rowSelection='single'
                 animateRows="true"
-                //style={{ width: '100%', height: '600px' }}
             />
+            <ConfirmDelete />
         </div>
     );
 }
